requirements.txt
streamlit>=1.32.0
openai>=1.0.0
akshare>=1.14.0
plotly>=5.18.0
pandas>=2.0.0
python-dotenv>=1.0.0
chromadb>=0.5.0
